Added proper default method to the Shape struct, which now renders to the canvas. Plus, new cleanup round.

+++ src/components/vectris/model/cell.rs
@@ -11,12 +11,12 @@ pub enum Color {
 
 #[derive(Copy, Clone, Debug, Default, PartialEq)]
 pub struct Cell {
-    pub coordinates: (u8, u8),
+    pub coordinates: (usize, usize),
     pub color: Option<Color>,
 }
 
 impl Cell {
-    pub fn new(x: u8, y: u8, color: Option<Color>) -> Self {
+    pub fn new(x: usize, y: usize, color: Option<Color>) -> Self {
         Self {
             coordinates: (x, y),
             color,
